sxhkd: switch to using Firefox instead of Chromium
I got a bit tired of Chromium being so slow sometimes, so I have been
using Firefox for a little while now. It works good enough, so I will
keep it (since Chromium is a bit slow, and I trust Mozilla more than
Google when it comes to spying on me).