alacritty + emacs + tmux: enable 24 bit colors everywhere
This seems to work at least. The colors in Emacs are a little
different, since the e.g. the `red` color is now actually red, and not
the `red` color from the base16 theme. There is probably a workaround
for this (e.g. using the `base08` color instead), but for now I don't
mind too much.