login + shell: begin moving general login configuration to `login`
The reason that this lived in shell was that I previously always used
a TTY for logging in, and thus the shell became the natural place for
putting this type of configuration. However, when using a display
manager the shell is not used for login, so configuration that is not
actually specific to the shell, but was needed when logging in, was
not loaded. Therefore, moving it to a general `login` directory seems
natural.