ansible: begin refactoring by moving `wallpaper` and `xorg`
This is the first step towards implementing my dotfiles as a big
Ansible playbook. It took a little while to figure out the structure
of this, but I am happy with the result. Now that I have taken the
first step, it should be a little bit easier porting all the other
configuration over to Ansible as well.

The biggest conceptual changes are the following:

* configuration files are deployed as files or templates, rather than
  as symlinks.
* The concept of "mixins", which allow roles to add functionality to
  each other, is introduced.

The former is to be certain that changes to e.g. paths where files are
stored have effect immediately. It centralizes those changes to
Ansible variables who will (hopefully) form a kind of dependency tree,
so that updating one variable and then re-deploying the playbook
should "just work".

The latter is to allow different parts of my dotfiles to live in their
own Ansible role, while still being loaded by other parts of my
configuration. An example of this is the setting of the wallpaper. It
should be done when X starts up, so it should be done in
`.xprofile`. However, this is the only relation the wallpaper has to X
configuration. Therefore, by placing a small script that only sets the
wallpaper in the `.xprofile` mixins, I can separate the setting of the
wallpaper from other X configuration. I expect to repeat this concept
throughout my dotfiles more and more, as I port the other
configurations over to Ansible.