ansible/alacritty: move `alacritty` configuration to Ansible role
This commit is fairly large, but most of it consists of splitting
things up into Ansible task/vars files, and moving stuff around.

The Ansible role tasks take care of
* installing `alacritty` and `alacritty-terminfo`;
* compiling my custom terminfos for 24-bit color support; and
* concatenating the colorscheme with my personal config.

The most notable change is that the script that previously
concatenated the colorscheme and my personal configuration is
deleted. This means that the concatenation is no longer done each time
`alacritty` is launched.