ansible/emacs: move `emacs` configuration to an Ansible role
This was easier than expected. Since we only need to symlink the
configuration directory, all that has to be done is to install
`emacs` (and, more specifically, the `emacs-nox` package to install
without X support) and create the symlink.

One could argue that all plugins should be installed here as
well. However, that seems a bit messy to do in Ansible, since it would
require running Emacs commands in Ansible.