macos/envensure: add role `envensure`
This role will be used to allow other roles to add environment
variables. For example, a Go role would like to set $GOPATH and
perhaps other variables, such as $GOBIN. It can then add an entry to
the `entries.txt` file using the `lineinfile` Ansible builtin, and a
shell can source `envensure.sh` to make sure it is included in the
environment.