README: remove `collection` from `ansible-galaxy` command
My reading of the documentation and the format of the file makes me
think that `ansible-galaxy` will figure out which are roles and which
are collections, and install them accordingly. On the other hand, I am
not sure if doing `ansible-galaxy collection install` will do the
right thing when installing "naked" roles.