ansible/alacritty: allow switching between colorschemes
Prior to this commit, if the base16 colorscheme was changed, there
would be multiple colorscheme files in the fragments directory. As
such, several colorschemes would be included in the generated
`alacritty.yml` file, and it would be arbitrary which won. With this
setup, it should be more resilient to change the colorscheme without
having to faff around with the fragments directory. Of course, all
problems would be fixed by just `rm -r ~/.config/alacritty`, but
having to remember that is annoying.