git: introduce a GitHub-flavored .gitconfig file.
The intent of this file is to allow me to work on the same repository in both
public and private fasions. This dotfiles repository is an example of that: I
maintain a private version of it, with a separate branch and a separate remote,
for settings specific to my current employer; but I also make commits to the
public version of it. For commits to the private branch I want other values for
`user.{name,email}`.