zsh: source environment files at the end of .zshrc in addition to .zshenv.
The reason is that macOS takes some liberties with modifying my $PATH variable
in a way that creates problems for me. Basically, despite my best efforts, it
reorders so that Homebrew paths are _still_ after system paths, even though I
use `brew shellenv` to set up it correctly in my environment files.

I'm not sure this is the neatest solution, but it works well and isn't _too_
hacky.