mercurial + python: define a `python` role and install `hg` (plus extensions) using that instead.
This hacky solution seems to work well enough, but it's absolutely a reaction to
the horrible package/dependency management that Python has. If this was all just
Go code, things would be so much easier... :)

The problem is primarily that because of the details about how Homebrew-provided
Python works, and how Homebrew's distribution of Mercurial depends on a
_different_ Python version than `brew install python3` gives, and how the
extensions I want are not distributed via Homebrew and has to be installed
through `pip` instead; but doing that the "default" way gives a huge warning
about breaking system packages because the Python installation is managed
externally, yada yada... it's a mess, and this is the solution I came up with.