roles/vscode: execute Go tests unconditionally.
Most of the time when I want to run a test from VSCode it is because I'm
actively working on it. When I suspect tests are flaky, I want to interactively
execute it a few times, and without `-count=1` the test will only be executed
the first time and then cached for subsequent runs. On paper this is a good
thing, since it avoids executing tests where nothing has changed, but in the
vast majority of cases things have changed (I'm in my editor writing code after
all), and this caching behavior creates a lot of annoyance when I do want to run
the same test multiple times.