spacemacs: add a basic Spacemacs configuration.
I struggled a bit to get this working properly along with OCaml, but it does the
job, well enough for now. I might come back and revisit and improve over time.

The `.spacemacs.env` file is probably a bit sensitive to minor changes in my
local environment, but like the comment says, it can be regenerated with `SPC f
e C-e`.